+++ app/src/androidTest/java/org/simple/clinic/di/TestAppModule.kt
@@ -27,7 +27,6 @@ import java.util.Locale
   TestStorageModule::class,
   TestClockModule::class,
   AppConfigModule::class,
-  TestPatientSearchModule::class,
   BruteForceProtectionModule::class,
   LoginModule::class,
   NetworkModule::class,