+++ app/src/main/java/org/simple/clinic/patient/DateOfBirth.kt
@@ -6,7 +6,6 @@ import org.simple.clinic.patient.DateOfBirth.Type.EXACT
 import org.simple.clinic.patient.DateOfBirth.Type.FROM_AGE
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.widgets.PatientSearchResultItemView
-import org.simple.clinic.widgets.PatientSearchResultItemView_Old
 import java.time.LocalDate
 import java.time.Period
 
@@ -51,13 +50,6 @@ data class DateOfBirth(
         userClock: UserClock
     ): DateOfBirth {
       return fromAgeOrDate(overdueAppointment.age, overdueAppointment.dateOfBirth, userClock)
-    }
-
-    fun fromPatientSearchResultViewModel_Old(
-        viewModel: PatientSearchResultItemView_Old.PatientSearchResultViewModel,
-        userClock: UserClock
-    ): DateOfBirth {
-      return fromAgeOrDate(viewModel.age, viewModel.dateOfBirth, userClock)
     }
 
     fun fromPatientSearchResultViewModel(