+++ app/src/main/java/org/simple/clinic/patient/filter/SearchPatientByName.kt
@@ -1,8 +1,0 @@
-package org.simple.clinic.patient.filter
-
-import org.simple.clinic.patient.PatientSearchResult
-import java.util.UUID
-
-interface SearchPatientByName {
-  fun search(searchTerm: String, names: List<PatientSearchResult.PatientNameAndId>): List<UUID>
-}