+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdEffect.kt
@@ -22,8 +22,6 @@ data class ParseScannedJson(val text: String) : ScanSimpleIdEffect()
 
 data class OpenPatientSummary(val patientId: UUID) : ScanSimpleIdEffect()
 
-data class OpenShortCodeSearch(val shortCode: String) : ScanSimpleIdEffect()
-
 data class OpenPatientSearch(
     val additionalIdentifier: Identifier?,
     val initialSearchQuery: String?,