+++ app/src/main/java/org/simple/clinic/shortcodesearchresult/UiActions.kt
@@ -1,8 +1,0 @@
-package org.simple.clinic.shortcodesearchresult
-
-import java.util.UUID
-
-interface UiActions {
-  fun openPatientSummary(patientUuid: UUID)
-  fun openPatientSearch()
-}